+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - `ZettlrFile` objects won't forcefully try to move a file to trash while handling watchdog events anymore.
 - `ZettlrRendererIPC` and `ZettlrIPC` now access the `ipc`-modules consistent with all other classes.
 - Generalised the `askFile()` function in `ZettlrWindow` for further purposes.
+- The `ZettlrConfig` now acts as an event emitter and emits `update`-events, whenever the configuration object changes. It can be subscribed to using `global.config.on` (to unsubscribe use `global.config.off`).
 
 # 0.20.0